Kalshi V2 read-side migration: restore position/order/fill legacy fields
The 2026-06 V2 wave that 410'd the order-mutation endpoints (fixed in
4789e47) also renamed the READ payload fields: positions now return
position_fp (no `position` key), orders return initial/fill/
remaining_count_fp (no count/filled_count/remaining_count), and orders/
fills report the YES-book leg in side/action (a buy-NO reads side=yes,
action=sell) with the outcome moved to outcome_side.

Every legacy read defaulted to zero silently, so all position-cap layers
in the bots saw held=0/resting=0 on every market in every run. The
per-market cap degenerated to a per-run counter. Post-mortem:
KXHIGHLAX-26JUL02-B72.5 filled 196.74 contracts against a 150 cap
(4 runs each granted full headroom; a 12:35Z taker sweep consumed three
19-lot rungs placed against 17 true contracts of headroom).

Client (KalshiClientsBaseV2ApiKey_FIXED): additive normalization in
get_positions/get_orders/get_order/get_fills — synthesize position/
count/filled_count/remaining_count as floats (contracts are fractional
now), rewrite side/action to the legacy customer view. V2 keys and
cursors preserved; V2-native readers unaffected. The customer-view
action rewrite also un-flips fills_cash in 90_ab_hi_no_report.sql,
which had been sign-inverted since the migration (KXHIGH_fills
self-heals on its next truncate-rebuild).

high_temp_trading: float-safe position/resting math; cancel sweep and
resting counts keyed on remaining_count > 0 (zeroed by Kalshi on
cancel/execution/expiry) instead of a status whitelist alone, so V2
status renames can't leak fills past the cap.

kxhigh_orderbook_logger: read V2 field names only — the legacy-first
fallbacks would have written synthesized cents into *_dollars FLOAT
columns (100x corruption) once normalization shipped.

mlb/nbamention/ncaab: round() before astype('Int64') — fractional
positions crash the strict cast. live_dashboard: keep fractional
positions instead of int-truncating |pos|<1 buckets to zero.

Verified: 27 offline tests against captured live payloads; live probe
through the patched client reads position=-196.74 on the post-mortem
ticker via the exact call path that returned 0 before.

+++ analysis/kxhigh/python/live_dashboard.py
@@ -914,7 +914,9 @@ def per_city_view(
                 "edge": float(r["edge"]) if pd.notna(r["edge"]) else None,
                 "yes_bid": int(r["ob_yes_bid"]) if pd.notna(r.get("ob_yes_bid")) else None,
                 "yes_offer": int(r["ob_yes_offer"]) if pd.notna(r.get("ob_yes_offer")) else None,
-                "position": int(pos_val) if pd.notna(pos_val) else 0,
+                # float, not int: fractional positions are real (e.g. -196.74)
+                # and int() truncation dropped |position|<1 buckets entirely
+                "position": float(pos_val) if pd.notna(pos_val) else 0.0,
                 "exposure_dollars": exposure_dollars,
                 "realized_pnl": realized_pnl,
                 "total_traded_dollars": float(r["total_traded_dollars"]) if pd.notna(r.get("total_traded_dollars")) else None,
